requirements.txt
streamlit
fastapi
uvicorn
tensorflow
faiss-cpu
joblib
numpy
pillow
scikit-learn
langchain
requests
